Root app seperated from main file.
