+++ lib/main.dart
@@ -1,20 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_ocr/view/login/view/login_view.dart';
-
-import 'core/init/navigation/navigation_routes.dart';
-import 'core/init/navigation/navigation_service.dart';
+import 'package:flutter_ocr/root_app.dart';
 
 Future<void> main() async {
   runApp(FlutterOcr());
-}
-
-class FlutterOcr extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: LoginView(),
-      navigatorKey: NavigationService.instance.navigatorKey,
-      onGenerateRoute: NavigationRoutes.instance.generateRoute,
-    );
-  }
 }
